+++ R/package.R
@@ -6,6 +6,7 @@
 #' @importFrom stats setNames
 #' @import methods
 #' @import formatters
+#' @import rlistings
 #' @import rtables
 #' @import officer
 #' @import flextable