+++ NEWS.md
@@ -1,4 +1,4 @@
-## rtables.officer 0.0.2.9004
+## rtables.officer 0.1.0
 
  * Added option to flush pages when exporting to `.docx`. It is possible to put tables on separate pages by doing `export_as_docx(add_page_break = TRUE)`.
  * Now it is possible to transform and export also `rlistings` objects.